+++ src/common/ainizeUtil.ts
