+++ src/common/util.ts
@@ -4,9 +4,9 @@ import Ainize from '@ainize-team/ainize-js';
 import { SetOperation, SetMultiOperation, TransactionInput } from '@ainblockchain/ain-js/lib/types';
 import Service from '@ainize-team/ainize-js/dist/service';
 
-import AinizeAuth from './common/ainize';
-import { SERVICE_NAME_MAP, MIN_GAS_PRICE } from './constants';
-import { HttpMethod, JobType, MessageMap } from './types';
+import AinizeAuth from './ainizeUtil';
+import { SERVICE_NAME_MAP, MIN_GAS_PRICE } from '../constants';
+import { HttpMethod, JobType, MessageMap } from '../types';
 
 export const buildData = (
   method: HttpMethod,
